wifi:  Add MinstrelHt rate control (based on patch originated by Ghada Badawy)
Minstrel-HT is a rate adaptation mechanism for 802.11n/ac standard based on
Minstrel [1].
The Minstrel-HT current design is based on the Linux implementation [2].
The Linux implementation is due to Felix Fietkau, first introduced in 2010 [3].

Minstrel-HT is designed for high-latency devices that implement a Multiple
Rate Retry (MRR) chain. This kind of device does not give feedback for every
frame retransmission, but just when a frame was correctly transmitted
(an ACK is received) or a frame transmission completely fails
(all retransmission attempts fail).

[1] https://wireless.wiki.kernel.org/en/developers/documentation/mac80211/ratecontrol/minstrel
[2] https://github.com/torvalds/linux/blob/master/net/mac80211/rc80211_minstrel_ht.c
[3] http://thread.gmane.org/gmane.linux.kernel.wireless.general/47372