internet: (fixes #231) Remove SocketAddressTag
SocketAddressTag was a long-standing API glitch. It was used to replicate the RecvFrom effect (i.e., to know the source address of packet) without calling RecvFrom.
This option is more harmful than useful, because in case of tunnels the new tag needs to replace the old one. Moreover, there is no real need to create a new API when there is a perfectly working one (i.e., RecvFrom).